% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-class.R
\docType{methods}
\name{coords}
\alias{coords}
\alias{coords}
\alias{coords,Heatmap-method}
\alias{coords<-}
\alias{coords<-,Heatmap-method}
\title{Return or set the coords in a Heatmap}
\usage{
coords(x)

\S4method{coords}{Heatmap}(x)

coords(x) <- value

\S4method{coords}{Heatmap}(x) <- value
}
\arguments{
\item{x}{A heatmap}

\item{value}{Replacement value}
}
\value{
integer, length 2, value of x@coords
}
\description{
Return or set the coords in a Heatmap
}
\examples{

data(HeatmapExamples)
coords(hm) = c(-100, 100)
}
