% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHeatmap.R
\name{default_color}
\alias{default_color}
\title{Predifined color palettes from RColorBrewer + Rainbow}
\usage{
default_color(col)
}
\arguments{
\item{col}{Character, RColorBrewer colorscheme or "Rainbow"

This function provides a convenient function to all color palettes from
RColorBrewer, and a better version of R's rainbow function (specifically
rev(rainbow(9, start=0, end=4/6)), so it starts blue and ends with red).}
}
\value{
character, a length-9 color palette
}
\description{
Predifined color palettes from RColorBrewer + Rainbow
}
\examples{
default_color("Blues")
default_color("Rainbow")
}
