% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{getScale}
\alias{getScale}
\title{Make an appropriate scale for a heatmap}
\usage{
getScale(x, y)
}
\arguments{
\item{x, }{y Min and max values for the heatmap

This function takes min/max values for a heatmap and
generates a scale either starting, ending or centered on
zero.}
}
\description{
Make an appropriate scale for a heatmap
}
\examples{
getScale(0.5, 5) # c(0, 5)
getScale(-6, -2) # c(-6, 6)
getScale(-6, 2) # c(-6, 6)
}

