% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{HeatmapExamples}
\alias{HeatmapExamples}
\alias{hm}
\alias{hm2}
\alias{mat}
\alias{rle_list}
\alias{string_set}
\alias{tata_pwm}
\alias{windows}
\title{Data for man page examples}
\format{An object of class \code{Heatmap} of length 500.}
\usage{
hm

hm2

mat

rle_list

string_set

tata_pwm

windows
}
\value{
invisible("HeatmapExamples")
}
\description{
Generated Data for examples

An example heatmap

A second example heatmap

An example matrix

An example RleList

An example DNAStringSet

An example PWM

An example GRanges
}
\keyword{datasets}
