% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHeatmap.R
\name{heatmapOptions}
\alias{heatmapOptions}
\title{Generate default options for a Heatmap}
\usage{
heatmapOptions(...)
}
\arguments{
\item{color}{A vector of colors (character)}

\item{box.width}{Numeric}

\item{x.ticks}{Logical}

\item{tcl}{Numeric}

\item{padj}{Numeric}

\item{cex.axis}{Numeric}

\item{scale}{Logical}

\item{scale.lwd}{Numeric}

\item{label.xpos}{Numeric}

\item{label.ypos}{Numeric}

\item{cex.label}{Numeric}

\item{label.col}{A Color (character)}

\item{legend}{Logical}

\item{legend.width}{Numeric}

\item{legend.pos}{'l' or 'r'}

\item{cex.legend}{Numeric}

\item{refline}{Logical}

\item{refline.width}{Numeric}

\item{transform}{Function or NA}

\item{plot.mai}{Numeric length 4}

\item{legend.mai}{Numeric length 4

Guide to Heatmap options

This is an reference to all the possible options for plotting heatmaps.
Some options are handled by heatmaps functions (either plotHeatmap or plotHeatmapList),
others are passed directly to plotting functions. Further explanation is available in the
vignette.

color: A vector of colors or a default color, see ?default_color. plotHeatmap will
       interpolate between these colors to form a scale.

box.width: The width of box around the heatmap, passed to box()

x.ticks: plot x axis ticks

tcl: Length of x axis ticks

padj: Vertical adjustment of x axis labels

cex.axis: cex for axis labels

scale: Plot scale or not

scale.lwd: Width for line around scale

label.xpos: x position for label, from left

label.ypos: y position for label, from top

cex.label: cex for axis labels

label.col: Color for label, white is often useful for dark plots

legend: Plot legend (scale indicating values for colors)

legend: Color for label, white is often useful for dark plots

legend.pos: Position of legend relative to heatmap: 'l' for left, 'r' for right

cex.legend: cex to use for legend marks

refline: Draw dashed line a coords = 0

refline.width: Width of reference line

transform: Transform values before plotting

plot.mai: Margins around plot

legend.mai: Margins around legend}
}
\description{
Generate default options for a Heatmap
}
\examples{

myOptions = heatmapOptions()
myOptions$color = "Reds"
# plotHeatmap(hm, options=myOptions)
}
\seealso{
plotHeatmap plotHeatmapList
}

