% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-class.R
\docType{methods}
\name{label}
\alias{label}
\alias{label}
\alias{label,Heatmap-method}
\alias{label<-}
\alias{label<-,Heatmap-method}
\title{Return or set the label in a Heatmap}
\usage{
label(x)

\S4method{label}{Heatmap}(x)

label(x) <- value

\S4method{label}{Heatmap}(x) <- value
}
\arguments{
\item{x}{A heatmap}

\item{value}{Replacement value}
}
\value{
character, value of hm@label
}
\description{
Return or set the label in a Heatmap
}
\examples{

data(HeatmapExamples)
label(hm) = "NewLabel"
label(hm) # "NewLabel"
}
