% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-class.R
\docType{methods}
\name{metadata}
\alias{metadata}
\alias{metadata}
\alias{metadata,Heatmap-method}
\alias{metadata<-}
\alias{metadata<-,Heatmap-method}
\title{Return or set the metadata in a Heatmap}
\usage{
metadata(x)

\S4method{metadata}{Heatmap}(x)

metadata(x) <- value

\S4method{metadata}{Heatmap}(x) <- value
}
\arguments{
\item{x}{A heatmap}

\item{value}{Replacement value}
}
\value{
list, value of hm@metadata
}
\description{
Store arbitrary metadata in a list, if desired.
}
\examples{

data(HeatmapExamples)
metadata(hm) = list(replicate=1, cell_line="ESC")
metadata(hm)$replicate == 1
}
