% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-class.R
\docType{methods}
\name{mirror}
\alias{mirror}
\alias{mirror,Heatmap-method}
\title{Reflect a heatmap in the y axis}
\usage{
mirror(x)

\S4method{mirror}{Heatmap}(x)
}
\arguments{
\item{x}{A heatmap}
}
\description{
Reflect a heatmap in the y axis
}
\section{Methods (by class)}{
\itemize{
\item \code{Heatmap}: Heatmap method
}}

