% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-class.R
\docType{methods}
\name{nseq}
\alias{nseq}
\alias{nseq}
\alias{nseq,Heatmap-method}
\alias{nseq<-}
\alias{nseq<-,Heatmap-method}
\title{Return or set nseq in a Heatmap}
\usage{
nseq(x)

\S4method{nseq}{Heatmap}(x)

nseq(x) <- value

\S4method{nseq}{Heatmap}(x) <- value
}
\arguments{
\item{x}{A heatmap}

\item{value}{Replacement value}
}
\value{
integer, value of hm@nseq
}
\description{
Return or set nseq in a Heatmap
}
\examples{

data(HeatmapExamples)
nseq(hm) = 1000
}
