% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHeatmapList.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot partition in a separate panel}
\usage{
plot_clusters(options)
}
\arguments{
\item{options}{heatmapOptions passed as a list

Two heatmapOptions values are relevant:

* partition Numeric vector containing relative sizes of the clusters
* colors Colors to use for clusters, additional colors are discarded

This function plots a vertical color scale (or legend). With the default parameters,
it looks good at about 1/5 the width of a heatmap, about 1cm x 10cm. This
function only plots the legend, it does not set margin parameters.}
}
\value{
invisible(0)
}
\description{
Plot partition in a separate panel
}
\examples{
data(HeatmapExamples)
opts = heatmapOptions()
opts$partition = c(1,2,3,4)
par(mai=opts$legend.mai)
plot_clusters(opts)
}
\seealso{
plotHeatmapList
}
