% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-class.R
\docType{methods}
\name{rev,Heatmap-method}
\alias{rev,Heatmap-method}
\title{Reflect a heatmap in the x axis}
\usage{
\S4method{rev}{Heatmap}(x)
}
\arguments{
\item{x}{A heatmap}
}
\value{
A heatmap
}
\description{
Reflect a heatmap in the x axis
}
