% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-class.R
\docType{methods}
\name{scale}
\alias{scale}
\alias{scale,Heatmap-method}
\alias{scale<-}
\alias{scale<-,Heatmap-method}
\title{Return or set the scale in a Heatmap}
\usage{
scale(x)

\S4method{scale}{Heatmap}(x)

scale(x, ...) <- value

\S4method{scale}{Heatmap}(x) <- value
}
\description{
Return or set the scale in a Heatmap
}
\examples{

data(HeatmapExamples)
scale(hm) = c(-1000, 1000)
}

