\name{seqPattern-package}
\alias{seqPattern-package}
\alias{seqPattern}
\docType{package}
\title{Visualising oligonucleotide patterns and motif occurences across a set
of ordered sequences}
\description{
Visualising oligonucleotide patterns and sequence motifs occurrences across a 
large set of sequences centred at a common reference point and sorted by a user 
defined feature.
}
\details{
\tabular{ll}{
Package: \tab SeqPattern\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-05\cr
License: \tab GPL-3\cr
Depends: \tab R (>= 3.0.1), methods\cr
}
}
\author{
Vanja Haberle

Maintainer: Vanja Haberle <vanja.haberle@gmail.com>
}
\keyword{ package }
