% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-class.R
\docType{methods}
\name{width,Heatmap-method}
\alias{width,Heatmap-method}
\title{Return the width of sequence represented in a heatmap}
\usage{
\S4method{width}{Heatmap}(x)
}
\arguments{
\item{x}{A heatmap}
}
\value{
integer
}
\description{
Return the width of sequence represented in a heatmap
}
