% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-class.R
\docType{methods}
\name{xm}
\alias{xm}
\alias{xm,Heatmap-method}
\title{Generate co-ordinates for each row of the image matrix of a Heatmap}
\usage{
xm(x)

\S4method{xm}{Heatmap}(x)
}
\arguments{
\item{x}{A Heatmap}
}
\description{
Generate co-ordinates for each row of the image matrix of a Heatmap
}
\section{Methods (by class)}{
\itemize{
\item \code{Heatmap}: Generate co-ordinates for each frow of the image matrix of a Heatmap
}}

