% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-class.R
\docType{methods}
\name{ym}
\alias{ym}
\alias{ym,Heatmap-method}
\title{Generate co-ordinates for each column of the image matrix of a Heatmap}
\usage{
ym(x)

\S4method{ym}{Heatmap}(x)
}
\arguments{
\item{x}{A Heatmap}
}
\value{
numeric, a list of co-ordinates for plotting values in hm@image
}
\description{
Generate co-ordinates for each column of the image matrix of a Heatmap
}
\section{Methods (by class)}{
\itemize{
\item \code{Heatmap}: Generate co-ordinates for each column of the matrix
}}

\examples{

data(HeatmapExamples)
ym(hm)
}
