\name{zebrafishPromoters}
\docType{data}
\alias{zebrafishPromoters}
\title{Zebrafish promoters sequences}
\description{
This is a \code{\link{DNAStringSet}} object that contains sequence of 1000 
randomly selected promoters active in 24 hpf zebrafish \emph{Danio rerio} 
embryos. The data is taken from Nepal \emph{et al.} Genome Research 2013, and 
represents regions flanking 400 bp upstream and 600 bp downstream of the 
dominant TSS detected by Cap analysis of gene expression (CAGE). It is intended 
to be used as an input data in running examples from \code{\link{seqPattern}} 
package help pages.
}
\usage{data(zebrafishPromoters)}
\value{A \code{\link{DNAStringSet}} object}
\format{A \code{\link{DNAStringSet}} object}
\keyword{datasets}
\references{
Nepal \emph{et al.} (2013) Dynamic regulation of the transcription initiation 
landscape at single nucleotide resolution during vertebrate embryogenesis, 
\emph{Genome Research} \bold{23}(11):1938-1950.
}