% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-class.R
\docType{class}
\name{Heatmap-class}
\alias{Heatmap-class}
\alias{length,Heatmap-method}
\alias{width,Heatmap-method}
\title{An S4 class to represent a heatmap}
\usage{
\S4method{length}{Heatmap}(x)

\S4method{width}{Heatmap}(x)
}
\description{
An S4 class to represent a heatmap
}
\section{Methods (by generic)}{
\itemize{
\item \code{length}: Return the number of sequences in a heatmap

\item \code{width}: Return the width of sequence represented in a heatmap
}}
\section{Slots}{

\describe{
\item{\code{image}}{A numeric Matrix}

\item{\code{scale}}{A length-2 vector}

\item{\code{coords}}{A length-2 integer vector}

\item{\code{nseq}}{An integer}

\item{\code{label}}{A character vector}

\item{\code{metadata}}{A list containing arbitrary metadata

A class used to represent a heatmap in a simple, self-contained way}
}}
\seealso{
CoverageHeatmap PatternHeatmap plotHeatmap plotHeatmapMeta
}

