% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-class.R
\name{Heatmap}
\alias{Heatmap}
\title{Function to create a heatmap object}
\usage{
Heatmap(mat, coords = NULL, label = "", nseq = NULL, scale = NULL,
  metadata = list())
}
\arguments{
\item{coords}{A length-2 integer vector}

\item{label}{A character vector}

\item{nseq}{An integer}

\item{scale}{A length-2 vector}

\item{metadata}{A list containing arbitrary metadata

Using this function avoids calling 'new' directly or manually setting coords and
nseq to integers. Other constructors exist for creating heatmaps from data, rather
than a raw matrix.}

\item{image}{A numeric Matrix}
}
\description{
Function to create a heatmap object
}
\examples{

library(HeatmapExamples)
hm = Heatmap(example_matrix, coords=c(-500, 500), label="Test")
}
\seealso{
PatternHeatmap CoverageHeatmap PWMScanHeatmap
}

