% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-class.R
\docType{methods}
\name{image}
\alias{image}
\alias{image,Heatmap-method}
\alias{image<-}
\alias{image<-,Heatmap-method}
\title{Return or set the image in a Heatmap}
\usage{
\S4method{image}{Heatmap}(x)

image(x) <- value

\S4method{image}{Heatmap}(x) <- value
}
\arguments{
\item{x}{A heatmap}

\item{value}{Replacement value}
}
\value{
matrix, from hm@image
}
\description{
Return or set the image in a Heatmap
}
\examples{

data(HeatmapExamples)
image(hm) = log(image(hm))
scale(hm) = c(0, max(image(hm)))
}
