% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap-class.R
\docType{methods}
\name{length,Heatmap-method}
\alias{length,Heatmap-method}
\title{Return the number of sequences in a heatmap}
\usage{
\S4method{length}{Heatmap}(x)
}
\arguments{
\item{x}{A heatmap}
}
\value{
integer, value of x@nseq
}
\description{
Return the number of sequences in a heatmap
}
